requirements.txt
flask
flask-cors
requests
beautifulsoup4
newspaper3k
lxml_html_clean
nltk
python-dotenv
openai
gunicorn
lxml
lxml_html_clean
